# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kindue household bill coordination API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /me/notification-settings:
    get:
      operationId: getNotificationSettings
      tags: [notifications]
      summary: Get notification settings for current user
      responses:
        "200":
          description: Notification settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
    patch:
      operationId: updateNotificationSettings
      tags: [notifications]
      summary: Update notification settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSettings"
  /me/push-tokens:
    post:
      operationId: registerPushToken
      tags: [notifications]
      summary: Register a push notification token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterPushTokenBody"
      responses:
        "201":
          description: Token registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PushToken"
  /me/push-tokens/{token}:
    delete:
      operationId: unregisterPushToken
      tags: [notifications]
      summary: Unregister a push notification token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Token removed
  /me/re-authenticate:
    post:
      operationId: reAuthenticate
      tags: [auth]
      summary: Re-authenticate before sensitive actions (role changes, account unlinking)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReAuthenticateBody"
      responses:
        "200":
          description: Re-authentication successful, returns a short-lived sensitive-action token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReAuthResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── HOUSEHOLDS ───────────────────────────────────────────────────────────────
...
  /households/{householdId}/bills:
    get:
      operationId: listBills
      tags: [bills]
      summary: List bills in a household
      parameters:
        - $ref: "#/components/parameters/householdId"
        - name: status
          in: query
          schema:
            type: string
            enum: [due, overdue, paid, pending_approval, all]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bills
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bill"
    post:
      operationId: createBill
      tags: [bills]
      summary: Add a bill (any member; CG/Other bills go to pending_approval)
      parameters:
        - $ref: "#/components/parameters/householdId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillBody"
      responses:
        "201":
          description: Created bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /households/{householdId}/bills/{billId}:
    get:
      operationId: getBill
      tags: [bills]
      summary: Get a bill
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/billId"
      responses:
        "200":
          description: Bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBill
      tags: [bills]
      summary: Update a bill (PU/Trustee only)
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/billId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBillBody"
      responses:
        "200":
          description: Updated bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
    delete:
      operationId: deleteBill
      tags: [bills]
      summary: Delete a bill (PU/Trustee only)
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/billId"
      responses:
        "204":
          description: Deleted
  /households/{householdId}/bills/{billId}/approve:
    post:
      operationId: approveBill
      tags: [bills]
      summary: Approve a pending bill (Trustee only)
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/billId"
      responses:
        "200":
          description: Approved bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /households/{householdId}/bills/{billId}/reject:
    post:
      operationId: rejectBill
      tags: [bills]
      summary: Reject a pending bill (Trustee only)
      parameters:
        - $ref: "#/components/parameters/householdId"
        - $ref: "#/components/parameters/billId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectBillBody"
      responses:
        "200":
          description: Rejected bill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
...
components:
  parameters:
    householdId:
      name: householdId
      in: path
      required: true
      schema:
        type: integer
...
    billId:
      name: billId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ── USER ────────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - email
        - createdAt
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
    ReAuthenticateBody:
      type: object
      properties:
        otpCode:
          type: string
          description: OTP code sent to user's email or phone
      required:
        - otpCode
    ReAuthResponse:
      type: object
      properties:
        sensitiveToken:
          type: string
          description: Short-lived token (15 min) required for sensitive operations
        expiresAt:
          type: string
          format: date-time
      required:
        - sensitiveToken
        - expiresAt

    # ── HOUSEHOLD ────────────────────────────────────────────────────────────────
...
    Bill:
      type: object
      properties:
        id:
          type: integer
        householdId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [utilities, telecom, insurance, mortgage_rent, auto_loan, credit_card, tolls, subscription, other]
        amount:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [due, overdue, paid, pending_approval, rejected]
        isRecurring:
          type: boolean
        recurrenceInterval:
          type: ["string", "null"]
          enum: [weekly, monthly, quarterly, annual, "null"]
        notes:
          type: ["string", "null"]
        billerName:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        requiresReceipt:
          type: boolean
          description: True if the next mark-paid action requires a receipt (for CG/Other)
        receiptUploaded:
          type: boolean
        riskFlags:
          type: array
          items:
            $ref: "#/components/schemas/RiskFlag"
        createdByRole:
          type: string
          enum: [primary_user, trustee, caregiver, other]
        assignedToMemberId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - householdId
        - name
        - category
        - status
        - isRecurring
        - requiresReceipt
        - receiptUploaded
        - riskFlags
        - createdByRole
        - createdAt
        - updatedAt
    CreateBillBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [utilities, telecom, insurance, mortgage_rent, auto_loan, credit_card, tolls, subscription, other]
        amount:
          type: number
        dueDate:
          type: string
          format: date-time
        isRecurring:
          type: boolean
        recurrenceInterval:
          type: string
          enum: [weekly, monthly, quarterly, annual]
        notes:
          type: string
        billerName:
          type: string
        assignedToMemberId:
          type: integer
      required:
        - name
        - category
    UpdateBillBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [utilities, telecom, insurance, mortgage_rent, auto_loan, credit_card, tolls, subscription, other]
        amount:
          type: number
        dueDate:
          type: string
          format: date-time
        isRecurring:
          type: boolean
        recurrenceInterval:
          type: string
          enum: [weekly, monthly, quarterly, annual]
        notes:
          type: string
        billerName:
          type: string
        assignedToMemberId:
          type: integer
    RejectBillBody:
      type: object
      properties:
        reason:
          type: string
      required:
        - reason
...
    RiskFlag:
      type: object
      properties:
        type:
          type: string
          enum: [overdue, shutoff_risk, late_fee_risk, low_balance_risk, upcoming_due]
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string
        daysUntilDue:
          type: ["integer", "null"]
      required:
        - type
        - severity
        - message

    # ── TRIAGE ───────────────────────────────────────────────────────────────────
...
    NotificationSettings:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        emailOverdue:
          type: boolean
        emailPendingApproval:
          type: boolean
        emailBillPaid:
          type: boolean
        emailBillRejected:
          type: boolean
        emailLowBalance:
          type: boolean
        pushOverdue:
          type: boolean
        pushPendingApproval:
          type: boolean
        pushBillPaid:
          type: boolean
        pushBillRejected:
          type: boolean
        pushLowBalance:
          type: boolean
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - emailOverdue
        - emailPendingApproval
        - emailBillPaid
        - emailBillRejected
        - emailLowBalance
        - pushOverdue
        - pushPendingApproval
        - pushBillPaid
        - pushBillRejected
        - pushLowBalance
        - updatedAt
    UpdateNotificationSettingsBody:
      type: object
      properties:
        emailOverdue:
          type: boolean
        emailPendingApproval:
          type: boolean
        emailBillPaid:
          type: boolean
        emailBillRejected:
          type: boolean
        emailLowBalance:
          type: boolean
        pushOverdue:
          type: boolean
        pushPendingApproval:
          type: boolean
        pushBillPaid:
          type: boolean
        pushBillRejected:
          type: boolean
        pushLowBalance:
          type: boolean
    PushToken:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        token:
          type: string
        platform:
          type: string
          enum: [ios, android, web]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - token
        - platform
        - createdAt
    RegisterPushTokenBody:
      type: object
      properties:
        token:
          type: string
        platform:
          type: string
          enum: [ios, android, web]
      required:
        - token
        - platform
...